Rename internal namespaces
Updates namespaces in Internal folder to include "Internal"
