+++ com.onesignal.unity.core/Runtime/Notifications/Internal/NotificationClickedResult.cs
@@ -26,8 +26,9 @@
  */
 
 using System;
+using OneSignalSDK.Notifications.Models;
 
-namespace OneSignalSDK.Notifications.Models {
+namespace OneSignalSDK.Notifications.Internal {
     /// <summary>
     /// See full documentation at
     /// https://documentation.onesignal.com/docs/sdk-notification-event-handlers#notification-opened-event
